requirements.txt
fastapi
uvicorn
requests
aiofiles
python-multipart
python-dotenv
